\name{e_dfsclass}
\alias{e_dfsclass}
\title{e_dfsclass}
\usage{
e_dfsclass()
}
\description{
Explaining datasets
}
\examples{

  df1 (df01_id_brand_conv)
  df2 (df02_study_exhaust)
  df3 (df03_gdp_local_tax)
  df4 (df04_consumption_gdp)
  df5 (df05_titanic)
  df6 (df06_mdis_house_finance18) 
  df7 (df07_caffe)
  df8 (df08_industry17) 
  df9 (df09_travel22) 
  df10 (df10_housing_status22)
  df11 (df11_mdis_donation18) 
  df12 (df12_survey_donation) 
  df13 (df13_final_exam)
  df14 (df14_final_exam)
  df15_suwon
  df16 (df16_corp2021)
  df17 (df17_OLS) 
  df18_anova
  df19_chisq
  df20_crispy 
  df21_foot 
  df22_path 
  df23_law

}
